+++ source/Public/Convert-PesterSyntax.ps1
@@ -43,7 +43,13 @@
         Get-ChildItem -Path "C:\Scripts" -Recurse -Filter "*.ps1" | Convert-PesterSyntax -Pester6
 
         Converts the syntax of all PowerShell files in the C:\Scripts directory and
-        its subdirectories to the default (newest) Pester syntax.
+        its subdirectories to Pester 6 syntax.
+
+    .EXAMPLE
+        Convert-PesterSyntax -Path (Get-ChildItem -Path "C:\Scripts" -Recurse -Filter "*.ps1") -Pester6
+
+        Converts the syntax of all PowerShell files in the C:\Scripts directory and
+        its subdirectories to Pester 6 syntax.
 #>
 function Convert-PesterSyntax
 {