+++ CHANGELOG.md
@@ -38,3 +38,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Convert-PesterSyntax`
   - The `Should` operators `BeLike` and `BeLikeExactly` was mistakenly not
     calling their respectively conversion function.
+  - Correctly handle abbreviated named parameters